Add zip files, of images, to the repository
The images of the zip files are the augmentations produced by the augmentation project, i.e., https://github.com/greyhypotheses/augmentation.  After running it in a cloud machine via a docker image, the produced augmentations, which are additionally zipped by the project, are downloaded.

Under normal/company conditions after the zip step the zip files would have been delivered to a Cloud repository.  In the absence of access to such resources, the files are downloaded via "scp", and committed to GitHub via this repository.